[ feature ] : 장소 상세 (#59)
* [ feature ] : 장소 상세 - domain layer (#46)

* [ feature ] : 장소 상세 - Data layer (#47)

* [ feature ] : 장소 상세 - data module

* [ feature ] : 장소 상세 - remote 모듈

* [ feature ] : 장소 상세 - Presentation layer (#48)

* [ feature ] : placeDetail 값 nullable로 변경

* [ feature ] : 장소 상세 - Presentation layer

* [ feature ] : 장소 상세 - App layer (#54)

* [ feature ] : 장소 상세 - App layer

* [ refactor ] : 사용하지않는 클래스 제거

* [ feature ] : 다크테마 제거하고 테마 및 상태바 색상 변경

+++ core-model/src/main/java/com/ddd/ansayo/core_model/place/Review.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class Review(
     @SerializedName("author_name") val authorName: String,
     val language: String,
-    val rating: Double,
+    val rating: Float,
     val text: String
 )
